refactor: [WLEO-556] Expose listeners (#17)
### Short description
This PR exposes the listener created by `addListener` in order to call
the `remove` method. It also removes the `removeListener` function as it
doesn't work as intended, thus the reference to the listener can be used
by calling the`remove` method instead.
<!--- Please always add a PR description as if nobody knows anything
about the context these changes come from. -->
<!--- Even if we are all from our internal team, we may not be on the
same page. -->
<!--- Write this PR as you were contributing to a public OSS project,
where nobody knows you and you have to earn their trust. -->
<!--- This will improve our projects in the long run! Thanks. -->

#### List of Changes
- Return the listener reference in `addListener`; 
- Remove the `removeListener` method;
- Update the documentation.
<!--- Describe your changes in detail -->

#### How Has This Been Tested?
This can be tested with the example app by starting and closing a
proximity flow.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, tests ran to see how
-->
<!--- your change affects other areas of the code, etc. -->

#### Screenshots (if appropriate):

<!--- Attach screenshots in case changes impact UI. -->