chore: [SIW-3109] Remove Android CRL validation (#18)
### Short description
This PR removes the Android CRL validation to allow offline
presentation.
<!--- Please always add a PR description as if nobody knows anything
about the context these changes come from. -->
<!--- Even if we are all from our internal team, we may not be on the
same page. -->
<!--- Write this PR as you were contributing to a public OSS project,
where nobody knows you and you have to earn their trust. -->
<!--- This will improve our projects in the long run! Thanks. -->

#### List of Changes
- Update the native proximity dependency on Android.
<!--- Describe your changes in detail -->

#### How Has This Been Tested?
This can be tested by putting the device offline and providing the
certificate to the `start` method of the example app. The
`isAuthenticated` flag should be `true` when using the appropriate
reader.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, tests ran to see how
-->
<!--- your change affects other areas of the code, etc. -->

#### Screenshots (if appropriate):

<!--- Attach screenshots in case changes impact UI. -->